fix(ENG-805): persisted `userId` not being shared between instances
a subtle bug that only became apparent when I started work on ENG-795
where I needed the `userId` to be persist between reloads.

Ultimately the issue was that the getter for the `userId` was being
called before the `consentLog` state had been read from the cookie
meaning that the test for `this.consentLogs[0]?.userId` would always be falsey
so a new user id would be generated.