fix!: leave it up to the consumer to initialise the client
BREAKING CHANGE: The client is no longer initialised automatically. The consumer must call `init` before using the client.
This is done automatically in the `ConsentProvider` component.

Until now we were hitting the network in the constructor which meant that we were making useless network requests
when the client was initialised on the server side.